Move to bloom/v1 package
