+++ pkg/storage/bloom/v1/tokenizer_test.go
@@ -1,4 +1,4 @@
-package bloomtokenizer
+package v1
 
 import (
 	"bufio"
